requirement.txt
mmcv==0.2.12
torch==1.1.0
editdistance
Polygon3
pyclipper
opencv-python==3.4.2.17
Cython
torchvision==0.3